app/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.42
psycopg2-binary==2.9.10
PyJWT==2.10.1
bcrypt==4.2.0
